+++ fabric/src/main/java/com/lx862/jcm/mod/block/StationNameStandingBlock.java
@@ -1,31 +1,23 @@
 package com.lx862.jcm.mod.block;
 
-import com.lx862.jcm.mod.block.base.Vertical3Block;
 import com.lx862.jcm.mod.block.entity.StationNameStandingBlockEntity;
 import org.mtr.mapping.holder.*;
 import org.mtr.mapping.mapper.BlockEntityExtension;
-import org.mtr.mapping.mapper.BlockWithEntity;
-import org.mtr.mod.block.IBlock;
+import org.mtr.mod.block.BlockStationNameTallStanding;
 
-public class StationNameStandingBlock extends Vertical3Block implements BlockWithEntity {
-    public StationNameStandingBlock(BlockSettings settings) {
-        super(settings);
+/**
+ * The old station name standing block. Block is moved to MTR Mod, now this is just a block wrapping the mtr mod variant to retain the Block Entity.
+ */
+@Deprecated
+public class StationNameStandingBlock extends BlockStationNameTallStanding {
+    public StationNameStandingBlock() {
+        super();
     }
 
     @Override
-    public VoxelShape getOutlineShape2(BlockState state, BlockView view, BlockPos pos, ShapeContext context) {
-        switch (IBlock.getStatePropertySafe(state, new Property<>(THIRD.data))) {
-            case LOWER:
-                VoxelShape vx1 = IBlock.getVoxelShapeByDirection(1, 0, 0, 2, 16, 1, IBlock.getStatePropertySafe(state, FACING));
-                VoxelShape vx2 = IBlock.getVoxelShapeByDirection(14, 0, 0, 15, 16, 1, IBlock.getStatePropertySafe(state, FACING));
-                return VoxelShapes.union(vx1, vx2);
-            case MIDDLE:
-                return IBlock.getVoxelShapeByDirection(1, 0, 0, 15, 16, 1, IBlock.getStatePropertySafe(state, FACING));
-            case UPPER:
-                return IBlock.getVoxelShapeByDirection(1, 0, 0, 15, 6, 1, IBlock.getStatePropertySafe(state, FACING));
-            default:
-                return VoxelShapes.empty();
-        }
+    public Item asItem2() {
+        // Redirect to official MTR block
+        return org.mtr.mod.Blocks.STATION_NAME_TALL_STANDING.get().asItem();
     }
 
     @Override