+++ CHANGELOG.md
@@ -1,10 +1,12 @@
-## JCM (v2.0.0-beta.14) for (MTR 4.0.0-prerelease.1) has been released!
+## JCM (v2.0.0-prerelease.1) for (MTR 4.0.0-prerelease.2) has been released!
 
 **Changes:**
 - To adapt to the new train driver key changes in MTR 4, **Operator Button** now allows choosing which driver keys are allowed.
 - - Configurable via brush
+- The JCM **Standing Station Name Block** has been moved to the official MTR Mod!
+- - The JCM variant will be kept for backward compatibility, but will no longer be obtainable in the inventory.
 
-**Fix:**
+**Fixes:**
 - Fix JSON PIDS ETA text potentially not following the destination's language
 - Fix Scripted PIDS Preset not working reliably
 - Fix some textures not being power-of-two resolution. The mipmap level can now be raised to 2 with solely MTR and JCM installed.